draft: add ThrottleRequestFilter for request rate limiting
A new ThrottleRequestFilter has been introduced to restrict the number of requests per minute that a client IP address can make to the server. This is implemented using the Caffeine library for caching and management of request counts. Changes also include adding the necessary Caffeine library dependency in the pom.xml file.