draft: remove Feed for user (and remove UserEntryInteraction)
The FeedService class has been updated to improve the deleteFeedForUser operation. Now, when a feed is removed from a user, all interactions that the user had with the entries of that feed are also deleted. This is achieved with the addition of a new 'UserEntryInteractionService'. This change helps keep the system clean by removing unneeded data, and increases efficiency when querying user interactions.