Refactor services and controllers to use DTOs everywhere
The modified the FeedController, EntryController, EntryService and FeedService methods to return DTOs (Data Transfer Objects) instead of full domain entities. This refactoring was done to isolate the internal representation of data from the data that is sent over the network, improving security by preventing sensitive data exposure. Now, the changes only return the necessary data to the client.