feat: implement search functionality for entries and feeds
This commit introduces a search functionality in the application. The new feature allows users to search for entries and feeds based on a search string. Queries were added to the Repository classes and new methods were added for handling the queries in the Services. A new REST endpoint inside a new SearchController is available for the client to utilize this functionality.