draft: add Mail Service and Mail Configuration
Added a MailService class to handle sending emails and a MailConfig class for configuring the mail system. Also, spring.mail properties in the application.properties file were defined to accommodate the new functionality. Furthermore, a new dependency, simple-java-mail, was added to the pom.xml file to aid with email operations.