draft: add email notifications on account request
In AuthService, a MailService was introduced that sends an email both to the user and admin when a new account is registered. Also, a new method 'findAllAdmins' was added in UserRepository that retrieves all with 'ROLE_ADMIN' role users. MailService utilizes the new functionality 'getAllAdmins' in UserService to retrieve all administrators' emails for sending notifications.