Implement encoded ID and user details retrieval from JWT token
We've made enhancements to handling JSON Web Tokens (JWT). Now, the user's ID is encoded and added as a claim in JWTs. Furthermore, we implemented a method to extract and decode this ID. In addition, we also created a method to load user details directly from the JWT, including the username and authorities. This provides a more efficient way of retrieving user details from tokens.